Move EncodingToTransportProcedures up to yarpc
