+++ v2/encoding.go
@@ -1,0 +1,85 @@
+// Copyright (c) 2018 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package yarpc
+
+import (
+	"context"
+	"fmt"
+)
+
+var _ UnaryTransportHandler = (*unaryTransportHandler)(nil)
+
+// EncodingToTransportProcedures converts encoding-level procedures to transport-level procedures.
+func EncodingToTransportProcedures(encodingProcedures []EncodingProcedure) ([]TransportProcedure, error) {
+	transportProcedures := make([]TransportProcedure, len(encodingProcedures))
+	for i, p := range encodingProcedures {
+		var spec TransportHandlerSpec
+		switch p.HandlerSpec.Type() {
+		case Unary:
+			spec = NewUnaryTransportHandlerSpec(&unaryTransportHandler{p})
+		default:
+			return nil, fmt.Errorf("unsupported handler spec type: %v", p.HandlerSpec.Type())
+		}
+
+		transportProcedures[i] = TransportProcedure{
+			Name:        p.Name,
+			Service:     p.Service,
+			HandlerSpec: spec,
+			Encoding:    p.Encoding,
+			Signature:   p.Signature,
+		}
+	}
+
+	return transportProcedures, nil
+}
+
+// Allows encoding-level procedures to act as transport-level procedures.
+type unaryTransportHandler struct {
+	h EncodingProcedure
+}
+
+func (u *unaryTransportHandler) Handle(ctx context.Context, req *Request, reqBuf *Buffer) (*Response, *Buffer, error) {
+	res := &Response{}
+	ctx, call := NewInboundCall(ctx)
+	if err := call.ReadFromRequest(req); err != nil {
+		return nil, nil, err
+	}
+
+	decodedBody, err := u.h.Codec.Decode(reqBuf)
+	if err != nil {
+		return res, nil, err
+	}
+
+	body, appErr := u.h.HandlerSpec.Unary().Handle(ctx, decodedBody)
+	call.WriteToResponse(res)
+
+	encodedBody, err := u.h.Codec.Encode(body)
+	if err != nil {
+		return res, nil, err
+	}
+
+	if appErr != nil {
+		res.ApplicationError = true
+		return res, encodedBody, appErr
+	}
+
+	return res, encodedBody, nil
+}
