+++ v2/yarpchttp/integration_test.go
@@ -33,7 +33,6 @@ import (
 	"go.uber.org/yarpc/v2"
 	"go.uber.org/yarpc/v2/yarpcerror"
 	"go.uber.org/yarpc/v2/yarpcjson"
-	"go.uber.org/yarpc/v2/yarpcrouter"
 	"go.uber.org/yarpc/v2/yarpctest"
 )
 
@@ -62,7 +61,7 @@ func TestBothResponseError(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(fmt.Sprintf("inbound(%v)-outbound(%v)", tt.inboundBothResponseError, tt.outboundBothResponseError), func(t *testing.T) {
-			procedures, err := yarpcrouter.EncodingToTransportProcedures(yarpcjson.Procedure("testFoo", testFooHandler))
+			procedures, err := yarpc.EncodingToTransportProcedures(yarpcjson.Procedure("testFoo", testFooHandler))
 			require.NoError(t, err)
 			doWithTestEnv(t, testEnvOptions{
 				Procedures: procedures,
