+++ v2/yarpcthrift/thriftrw-plugin-yarpc2/fx_test.go
@@ -86,7 +86,7 @@ func TestFxServer(t *testing.T) {
 			func() readonlystoreserver.Interface { return handler },
 			readonlystorefx.Server(),
 			func() jsonProcedures {
-				procedures, err := yarpcrouter.EncodingToTransportProcedures(
+				procedures, err := yarpc.EncodingToTransportProcedures(
 					yarpcjson.Procedure("echoJSON", echoJSON),
 				)
 				require.NoError(t, err)
